git/registerapi/user-management/src/main/java/com/example/user_management/service/UserService.java
package com.example.user_management.service;


import com.example.user_management.model.User;
import com.example.user_management.repository.UserRepository;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
public class UserService {

    @Autowired
    private UserRepository userRepository;
    @Autowired
    private EmailService emailService;

    public User createUser(User user) {
    	 // Save the user to the database
        User createdUser = userRepository.save(user);
    	 // Send a welcome email
        emailService.sendWelcomeEmail(createdUser.getEmail(), createdUser.getFirstName());
        return createdUser;
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User getUserById(Long id) {
        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
    }

    public User updateUser(Long id, User updatedUser) {
        User user = getUserById(id);
        user.setFirstName(updatedUser.getFirstName());
        user.setLastName(updatedUser.getLastName());
        user.setEmail(updatedUser.getEmail());
        return userRepository.save(user);
    }

    public void deleteUser(Long id) {
        User user = getUserById(id);
        user.setActive(false); // Soft delete
        userRepository.save(user);
    }
}

